+++ RoSA_Monitor/simulationwindow.cpp
@@ -6,12 +6,24 @@ SimulationWindow::SimulationWindow(QWidget *parent) :
     QMainWindow(parent),
     ui(new Ui::SimulationWindow), numOfTabs(1)
 {
+    ui->setupUi(this);
+
     for(int i = 0; i < LauncherManager::NUMBER_OF_LAUNCHERS; i++)
     {
         tab_Index[i] = 0;
         shellWindows[i] = nullptr;
     }
-    ui->setupUi(this);
+
+    ui->selectMapButton->setStyleSheet("QPushButton {"
+                                       "font-family: 'Times New Roman';"
+                                       "font-size: 18px;"
+                                       "font-weight: bold;"
+                                       "border-color: red;"
+                                       "background-color: lightcoral;"
+                                       "border-radius: 5px; "
+                                      "padding: 10px 20px; "
+                                       "color: black; "
+                                   "}");
 }
 
 SimulationWindow::~SimulationWindow()
@@ -317,6 +329,17 @@ void SimulationWindow::on_selectMapButton_clicked()
         manager.SetMap(map_name);
         QFileInfo mapInfo(*manager.GetMap());
         ui->mapLabel->setText(mapInfo.fileName());
+
+        ui->selectMapButton->setStyleSheet("QPushButton {"
+                                           "font-family: 'Times New Roman';"
+                                           "font-size: 18px; "
+                                           "font-weight: bold;"
+                                           "border: 2px solid #ccc;"
+                                           "border-radius: 5px;"
+                                          "padding: 10px 20px;"
+                                           "background: qlineargradient(x1:0, y1:0, x2:1, y2:1, stop:0 white, stop:1 #f0f0f0);"
+                                           "color: black; "
+                                       "}");
     }
 }